Update NavBar to tsx with appropriate TS
